collapse contracts decoders/ directory back to single decoders.rs
