--- src/contracts/common/decoders/mod.rs
@@ -1,4 +1,0 @@
-pub(crate) mod proto;
-mod text;
-
-pub(crate) use text::*;
